Add trash can to Java client, add integration tests

+++ services/repository/src/main/java/org/sagebionetworks/repo/web/service/TrashService.java
@@ -30,6 +30,6 @@ public interface TrashService {
 	/**
 	 * Retrieves entities (in the trash can) deleted by the specified user.
 	 */
-	PaginatedResults<TrashedEntity> viewTrash(String userId, Integer offset, Integer limit,
+	PaginatedResults<TrashedEntity> viewTrash(String userId, Long offset, Long limit,
 			HttpServletRequest request) throws DatastoreException, NotFoundException;
 }
